# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ConfigUpdateRequest:
    properties:
      config:
        type: object
    required:
    - config
    type: object
...
  dto.MenuUpdateRequest:
    properties:
      menus:
        type: object
    required:
    - menus
    type: object
...
  dto.SettingGroupResponse:
    properties:
      group:
        type: string
      settings:
        items:
          $ref: '#/definitions/dto.SettingItemResponse'
        type: array
    type: object
  dto.SettingItemResponse:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
...
  dto.ThemeMetaSyncRequest:
    properties:
      author:
        type: string
      description:
        type: string
      license:
        type: string
      name:
        type: string
      repo:
        type: string
      schema:
        type: object
      slug:
        type: string
      version:
        type: string
    required:
    - slug
    type: object
  dto.ThemeResponse:
    properties:
      author:
        type: string
      config:
        type: object
      description:
        type: string
      is_active:
        type: boolean
      license:
        type: string
      menus:
        type: object
      name:
        type: string
      repo:
        type: string
      schema:
        type: object
      slug:
        type: string
      version:
        type: string
    type: object
...
  dto.UpdateSettingGroupRequest:
    properties:
      settings:
        additionalProperties:
          type: string
        type: object
    required:
    - settings
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact:
    email: talen2004@163.com
    name: Talen
  description: 一个基于 Go 语言的现代化博客后端服务
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: FlecBlog
  version: v1
paths:
...
  /admin/settings/{group}:
    get:
      consumes:
      - application/json
      description: 获取指定分组的所有配置项
      parameters:
      - description: 配置分组
        enum:
        - basic
        - blog
        - notification
        - upload
        - ai
        - oauth
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.SettingGroupResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取配置分组
      tags:
      - 配置管理
    patch:
      consumes:
      - application/json
      description: 批量更新指定分组的配置项，限超级管理员
      parameters:
      - description: 配置分组
        enum:
        - basic
        - blog
        - notification
        - upload
        - ai
        - oauth
        in: path
        name: group
        required: true
        type: string
      - description: 配置更新内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSettingGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新配置分组
      tags:
      - 配置管理
  /admin/settings/ai/mcp-secret/reset:
    put:
      consumes:
      - application/json
      description: 重新生成并返回 MCP Secret，限超级管理员
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 重置 MCP Secret
      tags:
      - 配置管理
...
  /admin/themes:
    get:
      description: 获取全部主题实例
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ThemeResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 主题列表
      tags:
      - 主题管理
  /admin/themes/{slug}:
    get:
      description: 获取指定主题详情
      parameters:
      - description: 主题 slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 主题详情
      tags:
      - 主题管理
  /admin/themes/{slug}/config:
    put:
      consumes:
      - application/json
      description: 替换指定主题的 config JSON
      parameters:
      - description: 主题 slug
        in: path
        name: slug
        required: true
        type: string
      - description: 主题配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ConfigUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ThemeResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 更新主题配置
      tags:
      - 主题管理
  /admin/themes/{slug}/menus:
    put:
      consumes:
      - application/json
      description: 整体替换指定主题菜单
      parameters:
      - description: 主题 slug
        in: path
        name: slug
        required: true
        type: string
      - description: 主题菜单
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.MenuUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新主题菜单
      tags:
      - 主题管理
...
  /settings/{group}:
    get:
      consumes:
      - application/json
      description: 获取指定分组的公开配置项，根据配置项可见性规则筛选
      parameters:
      - description: 配置分组
        enum:
        - basic
        - blog
        - notification
        - upload
        - ai
        - oauth
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取配置分组
      tags:
      - 配置
...
  /themes/_sync:
    post:
      consumes:
      - application/json
      description: 主题镜像启动时上报元数据并激活当前主题
      parameters:
      - description: 主题元数据
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ThemeMetaSyncRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 同步主题元数据
      tags:
      - 主题
...
securityDefinitions:
  BearerAuth:
    description: 在请求头中添加 Bearer Token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"